+++ src/index.ts
@@ -2,6 +2,7 @@ export { IndigoDB } from "./indigodb";
 export { DataTypes } from "./dataTypes";
 export type { DataType } from "./dataTypes";
 export { BaseModel as Model } from "./models/baseModel";
+export type { AssociationOptions } from "./models/baseModel";
 export { DatabaseAdapter } from "./adapters/adapter";
 export type { TransactionContext } from "./adapters/adapter";
 export type { RealtimeGateway } from "./realtime/gateway";